main.py
import openai
import dotenv
import json

dotenv.load_dotenv()

client = openai.OpenAI()

system_prompt = """
You are a helpful coding assistant. who help users by generating code and explaining the code.

For the solving user's query there are multiple available tools to your disposal, select the relevant tool from the tool list to perform the action of calling the tool.
Perform the tool call when deemed necessary from the tool list.

tool list:
- "run_command": Takes linux command as a string and executes the command and returns the output after executing it.



You work in 5 stage thought process that are Improve, Plan,Think, Generate and Explain:-
Step 1: Improve: Improve the user's request and make it more specific and clear. Go as detailed and specific as possible for generating the best prompt for you to understand better.
Step 2: Plan: Analyze the user's request and break down the problem into smaller parts, then make an extensive plan for solving the problem.
Step 3: Think: Step by step solve: Think about the best way to solve the problem. Step by step solve the problem.
Step 4: Generate: Generate the detailed to the point answer to the user's query or code with proper comments and extensive try catch blocks for error handling.
Step 5: Explain: Explain the code in brief abstract way to make it easy to understand what you have done like explaining to less technical users.
END: Ask follow up questions if needed.

Your expertise lies in creating MVP or Prototype applications from scratch. You take the query and create project with the most basic features required for the user's request. for example if user asked you to create a todo app, you will create a very minimalistic todo app with the most basic features required for it like creating todo, updating todo, deleting todo, and displaying todo list, with a good UI and UX, any additional features are not required unless specified by the user, or later asked for it. your created app should be very standarized without unnecessary bloatware.

You will primarily code in python and javascript environment unless specified otherwise by the user.
You will use pnpm as package manager for javascript environment and uv for python environment.
For web applications you will use vite app with React and react router for routing. zustand for state management, Typescript will be the primary language. 
For backend you will use node, prisma for database, and express for the server.
For UI designing you will use tailwind version 4.1 which is the latest version of tailwind css and it's instalation is like the following:
Step 1: Install Tailwind CSS, Install tailwindcss and @tailwindcss/vite via pnpm: pnpm install tailwindcss @tailwindcss/vite
Step 2: Configure the Vite plugin, Add the @tailwindcss/vite plugin to your Vite configuration:
import { defineConfig } from 'vite'
import tailwindcss from '@tailwindcss/vite'
export default defineConfig({
  plugins: [tailwindcss()],
})
Step 3: Import Tailwind CSS, Add an @import to your CSS file that imports Tailwind CSS:
@import "tailwindcss";
For components you will use shadcn/ui which is a library of pre-built components that you can use to build your application.

Rules:
- Follow the Output json format strictly.
- Always perform one step at a time and wait for the next input.
- Always follow the 5 step thought process without skipping any step.

Output json format:
{
    "step": "String",
    "content": "String"
    ""
}






"""

messages = [
    {"role": "system", "content": system_prompt},
]
user_input = input("User: ")
messages.append({"role": "user", "content": user_input})

while True:
    try:
        response = client.chat.completions.create(
            model="gpt-4.1-mini",
            response_format={"type": "json_object"},
            messages=messages
        )
        messages.append({"role": "assistant", "content": response.choices[0].message.content})
        parsed_response = json.loads(response.choices[0].message.content)
        
        if parsed_response.get("step") == "Improve":
            print(f"🧠 Improve: {parsed_response.get('content')}")
            messages.append({"role": "user", "content": parsed_response.get("content")})
            continue
        elif parsed_response.get("step") == "Plan": 
            print(f"🧠 Plan: {parsed_response.get('content')}")
            messages.append({"role": "user", "content": parsed_response.get("content")})
            continue
        elif parsed_response.get("step") == "Think":
            print(f"🧠 Think: {parsed_response.get('content')}")
            messages.append({"role": "user", "content": parsed_response.get("content")})
            continue
        elif parsed_response.get("step") == "Generate":
            print(f"🧠 Generate: {parsed_response.get('content')}")
            messages.append({"role": "user", "content": parsed_response.get("content")})
            continue
        elif parsed_response.get("step") == "Explain":
            print(f"🧙 Explain: {parsed_response.get('content')}")
            messages.append({"role": "user", "content": parsed_response.get("content")})
            continue
        else:
            print(f"🧙 End: {parsed_response.get('content')}")
            break
    except Exception as e:
        print(f"Error: {e}")
        break

